Revert "cmd/gb: move env.go to cmd/gb"
This reverts commit d1b0828fa7dda68b82fddcc5c320f45466a94de2.
