+++ cmd/gb/env.go
@@ -1,33 +1,19 @@
 package main
 
-import (
-	"fmt"
-	"log"
-	"os"
-	"strings"
-)
+import "github.com/constabulary/gb/cmd"
 
-func mustGetwd() string {
-	wd, err := os.Getwd()
-	if err != nil {
-		log.Fatalf("unable to determine current working directory: %v", err)
-	}
-	return wd
+func init() {
+	registerCommand(EnvCmd)
 }
 
-// mergeEnv merges args into env, overwriting entries.
-func mergeEnv(env []string, args map[string]string) []string {
-	m := make(map[string]string)
-	for _, e := range env {
-		v := strings.SplitN(e, "=", 2)
-		m[v[0]] = v[1]
-	}
-	for k, v := range args {
-		m[k] = v
-	}
-	env = make([]string, 0, len(m))
-	for k, v := range m {
-		env = append(env, fmt.Sprintf("%s=%s", k, v))
-	}
-	return env
+var EnvCmd = &cmd.Command{
+	Name:      "env",
+	UsageLine: `env [var ...]`,
+	Short:     "print project environment variables",
+	Long: `
+Env prints project environment variables. If one or more variable names is 
+given as arguments, env prints the value of each named variable on its own line.
+`,
+	Run:           info,
+	SkipParseArgs: true,
 }
