+++ cmd/gb/generate.go
@@ -27,7 +27,7 @@ source files, for instance by running yacc.
 See 'go help generate'.
 `,
 	Run: func(ctx *gb.Context, args []string) error {
-		env := mergeEnv(os.Environ(), map[string]string{
+		env := cmd.MergeEnv(os.Environ(), map[string]string{
 			"GOPATH": fmt.Sprintf("%s%c%s", ctx.Projectdir(), filepath.ListSeparator, filepath.Join(ctx.Projectdir(), "vendor")),
 		})
